# openapi.yaml
openapi: 3.0.1
info:
  title: E-Commerce API
  version: 1.0.0
  description: A Basic CRUD API Example for an online store
paths:
  /api/user:
    summary: CRUD Operations for User Accounts
    get:
      summary: Gets the data of all users
      description: ''
      operationId: get_users
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The user ID.
                  first_name:
                    type: string
                    description: The user's first name.
                  last_Name:
                    type: string
                    description: The user's last name.
                  email:
                    type: string
                    description: The user's email address.
                  password:
                    type: string
                    description: The users password
        default:
          description: Default error sample response
      tags:
        - User
    post:
      summary: Create a new user
      description: Create a new user account
      operationId: create_user
      requestBody:
          description: A User's New Details, to be Updated.
          content:
            application/json: 
              schema:
                type: object
                properties:
                  first_name:
                    type: string
                    description: The user's first name.
                  last_Name:
                    type: string
                    description: The user's last name.
                  email:
                    type: string
                    description: The user's email address.
                  password:
                    type: string
                    description: The users password
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Success:
                  value: New User Record Successfully Added
        default:
          description: Default error sample response
      tags:
        - User
  /api/user/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: {}
    summary: Performing actions on specific user files
    description: Retreiving records based on id parameter
    get:
      summary: Gets a specific user
      description: Search for a specific user based on the id parameter
      operationId: get_specific_user
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The user ID.
                  first_name:
                    type: string
                    description: The user's first name.
                  last_Name:
                    type: string
                    description: The user's last name.
                  email:
                    type: string
                    description: The user's email address.
                  password:
                    type: string
                    description: The users password
                  created_at:
                    type: string 
                    description: Date the record was added
        default:
          description: Default error sample response
      tags:
        - User
    put:
      summary: Update a users record
      description: >-
        Provided details will be updated for the user matching the provided id
        parameter
      operationId: update_existing_user
      requestBody:
          description: A User's New Details, to be Updated.
          content:
            application/json: 
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The user ID.
                  first_name:
                    type: string
                    description: The user's first name.
                  last_Name:
                    type: string
                    description: The user's last name.
                  email:
                    type: string
                    description: The user's email address.
                  password:
                    type: string
                    description: The users password
                  created_at:
                    type: string 
                    description: Date the record was added
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Success:
                  value: User Record Successfully Updated
        default:
          description: Default error sample response
      tags:
        - User
    delete:
      summary: Delete a user
      description: Delete a User's record based on id parameter
      operationId: Delete_user
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Success:
                  value: User Record Successfully Deleted
        default:
          description: Default error sample response
      tags:
        - User
...
  /api/product:
    summary: CRUD operations for product items
    get:
      summary: Get all of the existing product details
      description: ''
      operationId: get_products
      responses:
        '200':
          description: A Product object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The user ID.
                  name:
                    type: string
                    description: The user's first name.
                  price:
                    type: integer
                    description: The user's last name.
                  category:
                    type: string
                    description: The user's email address.
                  stock:
                    type: integer
                    description: The users password
        default:
          description: Default error sample response
      tags:
        - Product
    post:
      summary: Create a new Product
      description: ''
      operationId: Create_Product
      requestBody:
          description: Adding a new Product
          content:
            application/json: 
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: The user's first name.
                  price:
                    type: integer
                    description: The user's last name.
                  category:
                    type: string
                    description: The user's email address.
                  stock:
                    type: integer
                    description: The users password
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Success:
                  value: New Product Record Successfully Added
        default:
          description: Default error sample response
      tags:
        - Product
  /api/product/{id}:
    summary: CRUD Operations for a specified product
    parameters:
      - in: path
        name: id
        required: true
        schema: {}
    get:
      summary: Get One Specified Product
      description: ''
      operationId: get_specific_product
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The user ID.
                  name:
                    type: string
                    description: The user's first name.
                  price:
                    type: integer
                    description: The user's last name.
                  category:
                    type: string
                    description: The user's email address.
                  stock:
                    type: integer
                    description: The users password
        default:
          description: Default error sample response
      tags:
        - Product
    put:
      summary: Update a Product's info
      description: ''
      operationId: update_product
      requestBody:
          description: Updating the details of an Existing Product
          content:
            application/json: 
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: The user's first name.
                  price:
                    type: integer
                    description: The user's last name.
                  category:
                    type: string
                    description: The user's email address.
                  stock:
                    type: integer
                    description: The users password
      responses:
        '200':
          description: A User object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The user ID.
                  name:
                    type: string
                    description: The user's first name.
                  price:
                    type: integer
                    description: The user's last name.
                  category:
                    type: string
                    description: The user's email address.
                  stock:
                    type: integer
                    description: The users password
        default:
          description: Default error sample response
      tags:
        - Product
    delete:
      summary: Delete a specified product
      description: ''
      operationId: delete_product
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Success:
                  value: Product Record Successfully Deleted
        default:
          description: Default error sample response
      tags:
        - Product
...